3回目修正

+++ todo/src/main/java/sample/app/dao/entity/Task.java
@@ -8,8 +8,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 public class Task {
-    private Integer id;
-    private Integer userId;
+    private Long id;
+    private Long userId;
 
     @NotBlank(message = "タイトルは必須です")
     private String title;
@@ -35,10 +35,10 @@ public class Task {
         return !endDate.isBefore(startDate);
     }
 
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
-    public Integer getUserId() { return userId; }
-    public void setUserId(Integer userId) { this.userId = userId; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
+    public Long getUserId() { return userId; }
+    public void setUserId(Long userId) { this.userId = userId; }
     public String getTitle() { return title; }
     public void setTitle(String title) { this.title = title; }
     public String getContent() { return content; }
